% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_db_connect.R
\name{access_db_connect}
\alias{access_db_connect}
\title{Connect to MS access database}
\usage{
access_db_connect(db_file_path)
}
\arguments{
\item{db_file_path}{path to a MS Access database}
}
\value{
\itemize{
\item A connection to a MS access database
}
}
\description{
This function is a wrapper for dbConnect and differs in that it takes a standard file
path as input.
}
