% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leadlag.R
\name{leadlag}
\alias{leadlag}
\title{Lead/Lag a logical}
\usage{
leadlag(lgl, bef = 1, aft = 1)
}
\arguments{
\item{lgl}{logical vector}

\item{bef}{integer, number of elements to lead by}

\item{aft}{integer, number of elements to lag by}
}
\value{
logical, same length as 'lgl'
}
\description{
Lead/Lag a logical
}
